[DEV][開發名稱: 範例程式] 1. 測試重新命名 2. 補完 webTestClient 全部測試

+++ src/main/java/tw/com/firstbank/fcbcore/fir/service/application/in/user/impl/DeleteUserUseCaseImpl.java
@@ -24,8 +24,12 @@ public class DeleteUserUseCaseImpl implements DeleteUserUseCaseApi {
     resp.setStatusCode(StatusCode.UNKNOWN_ERROR);
 
     try {
-      userService.deleteUser(mapper.toUserDto(requestCommand));
-      resp.setStatusCode(StatusCode.SUCCESS);
+      if(userService.hasUser(mapper.toUserDto(requestCommand))) {
+        userService.deleteUser(mapper.toUserDto(requestCommand));
+        resp.setStatusCode(StatusCode.SUCCESS);
+      } else {
+        resp.setStatusCode(StatusCode.DATA_NOT_FOUND);
+      }
     } catch (Exception ex) {
       log.error("Delete User Error.", ex);
     }
